Rename package to not use underscores
