+++ src/ext/java/org/opentripplanner/ext/updaterendpointfetcher/GBFSUpdaterEndpointFetcher.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.updater_endpoint_fetcher;
+package org.opentripplanner.ext.updaterendpointfetcher;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
