+++ src/main/java/org/opentripplanner/standalone/config/RouterConfig.java
@@ -2,7 +2,6 @@ package org.opentripplanner.standalone.config;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.MissingNode;
-import org.opentripplanner.ext.updater_endpoint_fetcher.GBFSUpdaterEndpointFetcher;
 import org.opentripplanner.routing.algorithm.raptor.transit.TransitTuningParameters;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.transit.raptor.api.request.RaptorTuningParameters;
